requirements.txt
openai>=1.0.0
requests>=2.28.1
beautifulsoup4>=4.11.1
pandas>=1.5.0
python-dotenv>=0.21.0
openpyxl>=3.0.10